requirements.txt
# Core
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0

# ML/Embeddings
cohere>=5.0.0
scikit-learn>=1.3.0

# Vector Database
pinecone-client>=3.0.0

# File Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# PDF Report Generation
weasyprint>=60.0
Jinja2>=3.1.0

# Data Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
tenacity>=8.2.0
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.0

# Development/Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
